add docker functionality

+++ server/api/routes/submit.js
@@ -3,17 +3,16 @@ const fs = require('fs');
 const router = express.Router();
 const { v4: uuidv4 } = require('uuid');
 const path = require('path');
-const { spawn } = require("child_process")
-const vm = require('vm');
-
+const exec = require('child_process').exec;
 const { models: { Test } } = require('../../db');
 
 router.post('/:id', async (req, res, next) => {
-
   // code is user code
   let code = req.body.code
+
   // create random filename
   let fileName = uuidv4() + '.js';
+
   // create filepath in temp folder utilizing random filename 
   let filePath = path.join(__dirname, `/temp/${fileName}`);
 
@@ -24,55 +23,60 @@ router.post('/:id', async (req, res, next) => {
         });
 
         let problem = response.dataValues.test;
- 
-        // create new temp file containing user code
-          fs.writeFileSync(filePath, code);
 
-          fs.appendFileSync(filePath, problem);
-        
-        // execute file in vm sandbox + capture return value of last line of executable code and assign to contextOutput
+        // create script to pass to docker container that contains code execution process
+        const runCode = `
+        const vm = require('vm');
+        try {
           const sandbox = {}
           vm.createContext(sandbox);
-          let script = new vm.Script(code + problem);
-          console.time('vm')
+          let script = new vm.Script(${JSON.stringify(code + problem)});
           const contextOutput = script.runInContext(sandbox, {
             console: console,
-        });
-          console.timeEnd('vm')
+          });
+          let consoleScript = new vm.Script(eval(${JSON.stringify(code)}));
+          console.log(contextOutput)
+        } catch(err) {
+        console.log(err)
+        }
+        `
 
-          console.time('eval')
-          eval(code)
-          console.timeEnd('eval')
-          let contextOutputArray = [contextOutput]
+        // create new temp file containing user code
+        fs.writeFile(filePath, runCode, (err) => {
+           if (err)
+           console.log(err);
+           else {
+            console.log("File written successfully\n");
 
-        // start stream to execute file to capture consoles or error
-          const consoleOutput = spawn("node", [filePath]);
-          let consoles = []
-
-        // create result object to send back to client
-          let result = { contextOutput: contextOutputArray, type: "", data: consoles | [], error: ""};
-        
-        // listen for consoles, add to console array to add to result object
-          consoleOutput.stdout.on('data', function (data) {
-            consoles.push(data.toString().trim());
-            result.type = "success";
-            result.data = consoles
-        });
-
-        // listen for errors, add to result object
-          consoleOutput.stderr.on('data', function (error) {
-            result.type = "error";
-            result.error = error.toString();
-        });
-        // listen for closing of stream, delete temp file + send result object to client
-          consoleOutput.on('close', (code) => {
+            exec(`docker run --rm -v ${filePath}:/runtest node:18.7.0 /bin/bash -c 'node runtest'`, 
+              (error, stdout, stderr) => {
+            if (error) {
+                console.log(`error: ${error.message}`);
+                fs.unlinkSync(filePath)
+                return;
+            }
+            if (stderr) {
+                console.log(`stderr: ${stderr}`);
+                fs.unlinkSync(filePath)
+                return;
+            }
+            stdout = stdout.split('\n')
+            stdout = stdout.filter(Boolean)
+            let results = {contextOutput: stdout[stdout.length - 1]}
+            stdout.pop();
+            results.consoleOutput = stdout
+            console.log(results)
+            res.json((results))
             fs.unlinkSync(filePath)
-            res.json(result)
-      });
+        });
+      }
+    });
 
     } catch(er){
       // if error, delete temp file, send error to client
-      fs.unlinkSync(filePath)
+      if (fs.existsSync(filePath)) {
+        fs.unlinkSync(filePath)
+      }
       res.json(String(er))
       next(er)
   }
